Tasks fix

+++ cyberlarp-api/src/main/java/com/example/cyberlarpapi/game/controllers/GameController.java
@@ -37,7 +37,7 @@ public class GameController {
         this.characterService = characterService;
     }
 
-    @Operation(summary = "Create a new game", description = "Create a new game in the system, providing name, description and game master user id")
+    @Operation(summary = "Create a new game [USER]", description = "Create a new game in the system, providing name, description and game master user id")
     @PostMapping
     public ResponseEntity<GameResponse> createGame(@RequestBody GameRequest request) {
         try {
@@ -59,60 +59,66 @@ public class GameController {
         }
     }
 
-    @Operation(summary = "Get all games", description = "Get all games in the system")
+    @Operation(summary = "Get all games [USER]", description = "Get all games in the system")
     @GetMapping
-    public ResponseEntity<List<GameResponse>> getAllGames() {
-        return ResponseEntity.ok(gameService.getAll().stream().map(GameResponse::new).toList());
+    public ResponseEntity<List<ShortGameResponse>> getAllGames() {
+        List<Game> games = gameService.getAll();
+        return ResponseEntity.ok(games.stream().map(ShortGameResponse::new).toList());
     }
 
 
-    @Operation(summary = "Get game by id", description = "Get game by id")
+    @Operation(summary = "Get game by id [GM]", description = "Get game by id")
     @GetMapping("/{id}")
     public ResponseEntity<GameResponse> getGameById(@PathVariable Integer id) {
         try {
-            return ResponseEntity.ok(new GameResponse(gameService.getById(id)));
+            _User user = userService.getCurrentUser();
+            Game game = gameService.getById(id);
+            if (!game.getGameMaster().getId().equals(user.getId())) {
+                return ResponseEntity.status(403).build();
+            }
+            return ResponseEntity.ok(new GameResponse(game));
         } catch (GameNotFoundException e) {
             return ResponseEntity.notFound().build();
+        } catch (UserServiceException e) {
+            return ResponseEntity.badRequest().build();
         }
     }
 
-    @Operation(summary = "Update game text data by id", description = "Update game text data by id, providing name and description")
+    @Operation(summary = "Update game [GM]", description = "Update game data by id, providing name and description")
     @PutMapping("/{id}/textData")
     public ResponseEntity<GameResponse> updateGameTextDataById(@PathVariable Integer id, @RequestBody GameRequest request) {
         try {
+            _User user = userService.getCurrentUser();
             Game game = gameService.getById(id);
+            if (!game.getGameMaster().getId().equals(user.getId())) {
+                return ResponseEntity.status(403).build();
+            }
             game.setName(request.getName());
             game.setDescription(request.getDescription());
             game = gameService.save(game);
             return ResponseEntity.ok(new GameResponse("Game updated successfully", game));
         } catch (GameNotFoundException e) {
             return ResponseEntity.notFound().build();
+        } catch (UserServiceException e) {
+            return ResponseEntity.badRequest().build();
         }
     }
 
-    @Operation(summary = "Delete game by id", description = "Delete game by id")
+    @Operation(summary = "Delete game [GM]", description = "Delete game by id")
     @DeleteMapping("/{id}")
     public ResponseEntity<GameResponse> deleteGame(@PathVariable Integer id) {
         try {
+            _User user = userService.getCurrentUser();
+            Game game = gameService.getById(id);
+            if (!game.getGameMaster().getId().equals(user.getId())) {
+                return ResponseEntity.status(403).build();
+            }
             gameService.deleteById(id);
             return ResponseEntity.ok(new GameResponse("Game deleted successfully"));
-        } catch (GameServiceException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Operation(summary = "Make user owner of game", description = "Make user owner of game by providing game id and user id")
-    @PutMapping("/{id}/gameMaster/{userId}")
-    public ResponseEntity<GameResponse> makeUserOwnerOfGame(@PathVariable Integer id, @PathVariable Integer userId) {
-        try {
-            Game game = gameService.getById(id);
-            _User user = userService.getUserById(userId);
-            gameService.makeUserOwnerOfGame(id, user);
-            return ResponseEntity.ok(new GameResponse("User is now the owner of the game", game));
+        } catch (GameServiceException | UserServiceException e) {
+            return ResponseEntity.badRequest().build();
         } catch (GameNotFoundException e) {
             return ResponseEntity.notFound().build();
-        } catch (UserServiceException e) {
-            throw new RuntimeException(e);
         }
     }
 
@@ -122,6 +128,21 @@ public class GameController {
     public static class GameRequest {
         private String name;
         private String description;
+    }
+
+    @Getter
+    public static class ShortGameResponse {
+        private final Integer id;
+        private final String name;
+        private final String description;
+        private final Integer gameMasterId;
+
+        public ShortGameResponse(Game game) {
+            this.id = game.getId();
+            this.name = game.getName();
+            this.description = game.getDescription();
+            this.gameMasterId = game.getGameMaster().getId();
+        }
     }
 
     @Getter
